Add files via upload
Vectors header file is used in this code because these are dynamic arrays, and it means these arrays can adjust their size automatically. And it is providing flexibility. One function is made for the merge sorting with the name of merge. And the second function is for dividing the array. Four threads are created to divide and sort the array and then all four threads are joined. In driver function, input from the user is taken.